Fix: Ensure input dtype matches weight dtype before linear operation
When mixed precision training is used, the input may arrive as FP32
while the weight is FP16. Convert input to match weight dtype before
applying the linear transformation to avoid dtype mismatch.

This handles the case where autocast or other mechanisms convert
the input to FP32 upstream.

+++ modules/dobi_matryoshka_svd.py
@@ -109,7 +109,8 @@ class DobiMatryoshkaSVDLayer(nn.Module):
         original_shape = x.shape
 
         # === STEP 1: Apply original weight (from Dobi-SVD) ===
-        # Keep in original dtype (FP16) to avoid dtype mismatch
+        # Ensure input matches weight dtype to avoid mismatch
+        x = x.to(self.ori.weight.dtype)
         x = self.ori(x)
 
         # === STEP 2: Convert to FP32 for SVD ===
